Extra CLI command component
