+++ app/pages/project/instances/instance/SerialConsolePage.tsx
@@ -1,17 +1,11 @@
-import { Suspense, lazy, useState } from 'react'
+import { Suspense, lazy } from 'react'
 import { Link } from 'react-router-dom'
 
 import { useApiQuery } from '@oxide/api'
-import {
-  Button,
-  Modal,
-  PrevArrow12Icon,
-  Spinner,
-  Success12Icon,
-  useTimeout,
-} from '@oxide/ui'
+import { Button, PrevArrow12Icon, Spinner } from '@oxide/ui'
 import { MiB } from '@oxide/util'
 
+import EquivalentCliCommand from 'app/components/EquivalentCliCommand'
 import { SerialConsoleStatusBadge } from 'app/components/StatusBadge'
 import { useInstanceSelector } from 'app/hooks'
 import { pb } from 'app/util/path-builder'
@@ -76,59 +70,6 @@ export function SerialConsolePage() {
         </div>
       </div>
     </div>
-  )
-}
-
-function EquivalentCliCommand({ command }: { command: string }) {
-  const [isOpen, setIsOpen] = useState(false)
-  const [hasCopied, setHasCopied] = useState(false)
-
-  function handleDismiss() {
-    setIsOpen(false)
-  }
-
-  useTimeout(() => setHasCopied(false), hasCopied ? 2000 : null)
-
-  const handleCopy = () => {
-    window.navigator.clipboard.writeText(command).then(() => {
-      setHasCopied(true)
-    })
-  }
-
-  return (
-    <>
-      <Button variant="ghost" size="sm" className="ml-2" onClick={() => setIsOpen(true)}>
-        Equivalent CLI Command
-      </Button>
-      <Modal isOpen={isOpen} onDismiss={handleDismiss}>
-        <Modal.Title>CLI Command</Modal.Title>
-
-        <Modal.Section>
-          {/* todo: fix the token to disable contextual alternates in the mono font */}
-          <pre className="flex w-full rounded border px-4 py-3 !normal-case !tracking-normal text-mono-md bg-default border-secondary [font-feature-settings:_'calt'_off]">
-            <div className="mr-2 select-none text-quaternary">$</div>
-            {command}
-          </pre>
-        </Modal.Section>
-        <Modal.Footer
-          onDismiss={handleDismiss}
-          onAction={handleCopy}
-          actionText={
-            <>
-              <span className={hasCopied ? 'invisible' : ''}>Copy command</span>
-              <span
-                className={`absolute left-1/2 top-1/2 flex -translate-x-1/2 -translate-y-1/2 items-center ${
-                  hasCopied ? '' : 'invisible'
-                }`}
-              >
-                <Success12Icon className="mr-2 text-accent" />
-                Copied
-              </span>
-            </>
-          }
-        />
-      </Modal>
-    </>
   )
 }
